### 2019-08-01 #1618 @whedon
[PRE REVIEW]: Pakman: a modular, efficient and portable tool for approximate Bayesian inference
**Submitting author:** @ThomasPak (<a href="http://orcid.org/0000-0002-7198-7688">Thomas Pak</a>)
**Repository:** <a href="https://github.com/ThomasPak/pakman" target ="_blank">https://github.com/ThomasPak/pakman</a>
**Version:** v1.0.0
