### 2019-08-01 #1618 @jedbrown
@ThomasPak Could you please address [Community Guidelines](https://joss.readthedocs.io/en/latest/review_criteria.html#community-guidelines)?

In your paper, this naming would result in `PacMan` (which of course is impossible to search for). Using your last name makes it distinguishable, but I think it's worth thinking about what that looks like to potential contributors if building a community is one of your goals (it need not be for JOSS, but many JOSS authors wish to).
![Screenshot_20190801_113522](https://user-images.githubusercontent.com/3303/62314666-c5c64700-b450-11e9-94ee-b5778dc3c518.png)
