### 2019-08-01 #1618 @ThomasPak
@jedbrown Hi Jed, thanks for your comment.

I designed Pakman so that it is easy to extend with additional ABC algorithms. This is mentioned in the [code documentation](https://codedocs.xyz/ThomasPak/pakman/index.html). In addition, I  wrote a page that goes into practical detail on how to do this (see [Implementing a Controller subclass](https://codedocs.xyz/ThomasPak/pakman/controller.html)).

I would want and expect that people contribute to this project, and I do not claim sole ownership of Pakman. The name was merely the result of a brainstorming session where I tried to find a name that is memorable and, as you said, makes it distinguishable.

I have to admit though that I did not give too much thought to how the name would be perceived by potential contributors. Would you suggest adding more emphasis on community building in the documentation or finding a new name altogether?
