### 2019-08-01 #1618 @jedbrown
I think the documentation on extensibility is nice, but it would help to add a [CONTRIBUTING.md](https://help.github.com/en/articles/setting-guidelines-for-repository-contributors) and explain how you would like for issues and questions to be raised. It's common to ask for pull requests and issues to be created on GitHub, but you might want something different or might want to provide further guidelines. There are many facets to creating a welcoming community, but expressing that contributions are encouraged and will be recognized, plus adopting a [code of conduct](https://help.github.com/en/articles/adding-a-code-of-conduct-to-your-project) are best practices.
Regarding the name, that's a big change, but it's easier to do now than later. It's entirely your choice, but I think it's worth thinking about how it will be perceived by a community. Linux is successful, but I can't think of many community projects named after an initial author that have thrived.
