### 2019-08-03 a8488a4
Enforce maximum line length of 79
