### 2019-08-03 abe5f9c
Add CONTRIBUTING.md and CODE_OF_CONDUCT.md
