### 2019-08-03 b2e6e4c
Link to CONTRIBUTING.md in README.md
