### 2019-08-03 #1618 @ThomasPak
@jedbrown Thanks for the resources on writing contributing guidelines. I have now added a [CONTRIBUTING.md](https://github.com/ThomasPak/pakman/blob/master/CONTRIBUTING.md), as well as a [CODE_OF_CONDUCT.md](https://github.com/ThomasPak/pakman/blob/master/CODE_OF_CONDUCT.md). Apologies for the delay, this is my first open source project so it is all a bit new to me.

After discussing the naming matter with my co-authors, we ultimately decided against changing the name.
