### 2019-08-12 #1618 @ThomasPak
Dear @jedbrown, I would like to suggest the following reviewers:

- fabian-s
- becarioprecario
- mattpitkin
- Volkerschmid
- gweodoo
- gonsie
- mikaem

I selected these reviewers based on their topic areas including either Bayesian statistics or MPI, as well as their knowledge of C++.

In addition, darrenjw is not on the suggested list of reviewers, but I believe he would be a suitable reviewer for this paper as well.
