### 2019-08-18 3807c8b
Implement option to write Pakman output to file
