### 2019-08-25 2418233
Refactor names of static class members
Static class members are now prefixed with `s_` instead of `m_` to
distinguish them from non-static class members.
