+++ src/master/MPIWorkerHandler.cc
@@ -10,19 +10,19 @@
 
 // Initialize static child communicator of MPIWorkerHandler to the
 // null communicators (MPI_COMM_NULL)
-MPI_Comm MPIWorkerHandler::m_child_comm = MPI_COMM_NULL;
+MPI_Comm MPIWorkerHandler::s_child_comm = MPI_COMM_NULL;
 
 MPIWorkerHandler::MPIWorkerHandler(const Command& simulator,
         const std::string& input_string) :
     AbstractWorkerHandler(simulator, input_string)
 {
     // Spawn  MPI child process if it has not yet been spawned
-    if (m_child_comm == MPI_COMM_NULL)
-        m_child_comm = spawn_worker(m_simulator);
+    if (s_child_comm == MPI_COMM_NULL)
+        s_child_comm = spawn_worker(m_simulator);
 
     // Write input string to spawned MPI process
     MPI_Send(input_string.c_str(), input_string.size() + 1, MPI_CHAR,
-            WORKER_RANK, MANAGER_MSG_TAG, m_child_comm);
+            WORKER_RANK, MANAGER_MSG_TAG, s_child_comm);
 }
 
 MPIWorkerHandler::~MPIWorkerHandler()
@@ -35,7 +35,7 @@ bool MPIWorkerHandler::isDone()
 {
     // Probe for result if result has not yet been received
     if (    !m_result_received &&
-            iprobe_wrapper(WORKER_RANK, WORKER_MSG_TAG, m_child_comm))
+            iprobe_wrapper(WORKER_RANK, WORKER_MSG_TAG, s_child_comm))
     {
         // Receive message
         m_output_buffer.assign(receiveMessage());
@@ -52,12 +52,12 @@ bool MPIWorkerHandler::isDone()
 
 std::string MPIWorkerHandler::receiveMessage() const
 {
-    return receive_string(m_child_comm, WORKER_RANK, WORKER_MSG_TAG);
+    return receive_string(s_child_comm, WORKER_RANK, WORKER_MSG_TAG);
 }
 
 int MPIWorkerHandler::receiveErrorCode() const
 {
-    return receive_integer(m_child_comm, WORKER_RANK, WORKER_ERROR_CODE_TAG);
+    return receive_integer(s_child_comm, WORKER_RANK, WORKER_ERROR_CODE_TAG);
 }
 
 void MPIWorkerHandler::discardResults()
@@ -68,7 +68,7 @@ void MPIWorkerHandler::discardResults()
     if (!m_result_received)
     {
         // Timeout if message is not ready yet
-        while (!iprobe_wrapper(WORKER_RANK, WORKER_MSG_TAG, m_child_comm))
+        while (!iprobe_wrapper(WORKER_RANK, WORKER_MSG_TAG, s_child_comm))
             std::this_thread::sleep_for(MAIN_TIMEOUT);
 
         // Receive message
@@ -87,16 +87,16 @@ void MPIWorkerHandler::terminateStatic()
     // If this function is called, the Worker must be in an idle state, so it
     // is not necessary to discard results from Worker.
 
-    // If m_child_comm is the null communicator, the Worker has already been
+    // If s_child_comm is the null communicator, the Worker has already been
     // terminated, so nothing needs to be done.
-    if (m_child_comm == MPI_COMM_NULL)
+    if (s_child_comm == MPI_COMM_NULL)
         return;
 
     // Else, send termination signal to Worker
     int signal = TERMINATE_WORKER_SIGNAL;
     MPI_Send(&signal, 1, MPI_INT, WORKER_RANK, MANAGER_SIGNAL_TAG,
-            m_child_comm);
+            s_child_comm);
 
[…]
