### 2019-08-25 1f61686
Prefix global variable with `g_`
