### 2019-08-25 2394ddd
Implement OutputStreamHandler to handle output
