+++ src/core/CMakeLists.txt
@@ -2,5 +2,6 @@ add_library (core
     Arguments.cc
     LongOptions.cc
     Command.cc
+    OutputStreamHandler.cc
     utils.cc
     )
