### 2019-09-02 #1618 @jedbrown
@Volkerschmid @gweodoo @gonsie :wave: Would you be available to review this submission for JOSS?
