### 2019-09-06 #1716 @jedbrown
@jmlarson1, @gonsie :wave: Welcome and thanks for agreeing to review! The comments from @whedon above outline the review process, which takes place in this thread (possibly with issues filed in the [Pakman repository](https://github.com/ThomasPak/pakman) and mentioning the URL of this issue). I'll be watching this thread if you have any questions.
