### 2019-09-10 d3f3a72
Use if conditional to write out comma
This fixes a bug that occurs where no output was written when
parameter_tokens was empty.  In the former implementation, the rewind
operation caused every newline to be overwritten by another newline.
