### 2019-09-10 #1716 @jmlarson1
Some things of note:

1. I received the following warning when trying to build.

````
CMake Warning at examples/epithelial-cell-growth/CMakeLists.txt:3 (message):
  CHASTE_DIR ("") is not a valid directory, cannot build epithelial cell
  growth example.
````

2. I tried to run the tests, but it hangs;
````
[jlarson@x1carbon build]$ ctest
Test project /home/jlarson/software/pakman/build
      Start  1: MPIMasterSweepStandardSimulatorMatch
 1/32 Test  #1: MPIMasterSweepStandardSimulatorMatch ..........   Passed    0.03 sec
      Start  2: MPIMasterSweepStandardSimulatorError
 2/32 Test  #2: MPIMasterSweepStandardSimulatorError ..........   Passed    0.02 sec
      Start  3: SerialMasterSweepStandardSimulatorMatch
 3/32 Test  #3: SerialMasterSweepStandardSimulatorMatch .......   Passed    0.01 sec
      Start  4: SerialMasterSweepStandardSimulatorError
 4/32 Test  #4: SerialMasterSweepStandardSimulatorError .......   Passed    0.01 sec
      Start  5: MPIMasterRejectionStandardSimulatorMatch
 5/32 Test  #5: MPIMasterRejectionStandardSimulatorMatch ......   Passed    0.04 sec
      Start  6: MPIMasterRejectionStandardSimulatorError
 6/32 Test  #6: MPIMasterRejectionStandardSimulatorError ......   Passed    0.02 sec
      Start  7: SerialMasterRejectionStandardSimulatorMatch
 7/32 Test  #7: SerialMasterRejectionStandardSimulatorMatch ...   Passed    0.02 sec
      Start  8: SerialMasterRejectionStandardSimulatorError
 8/32 Test  #8: SerialMasterRejectionStandardSimulatorError ...   Passed    0.00 sec
      Start  9: MPIMasterSMCStandardSimulatorMatch
 9/32 Test  #9: MPIMasterSMCStandardSimulatorMatch ............   Passed    0.20 sec
      Start 10: MPIMasterSMCStandardSimulatorError
10/32 Test #10: MPIMasterSMCStandardSimulatorError ............   Passed    0.02 sec
      Start 11: SerialMasterSMCStandardSimulatorMatch
11/32 Test #11: SerialMasterSMCStandardSimulatorMatch .........   Passed    0.13 sec
      Start 12: SerialMasterSMCStandardSimulatorError
12/32 Test #12: SerialMasterSMCStandardSimulatorError .........   Passed    0.00 sec
      Start 13: RunMPISimulatorMatch
````

3. Aren't the examples `prior-sampler.py` or `simulator.py` in `/pakman/examples/biased-coin-flip`
just wrappers around numpy? (And therefore not a good example of the code being used?)

4. The paper does not "clearly state what problems the software is designed to solve and who the target audience is?"

5. There are no references in the paper

6. The code does not seem to be well documented. Consider the example script:
https://github.com/ThomasPak/pakman/blob/master/examples/sis-model/simulator.py

The `run_SIS_simulation` routine does not state what the various arguments represent.

7. The paper does not sufficiently summarize the state of the field.
