### 2019-09-11 0d1fa18
Read input from file in run-mpi-simulator
Reading from standard input is not defined in the MPI standard and as a
result different MPI implementations have different ways of dealing with
it.  Therefore, I changed the specification of run-mpi-simulator so that
it reads its input from a file instead.
