+++ utils/run-mpi-simulator.cc
@@ -1,7 +1,9 @@
 #include <thread>
+#include <fstream>
 #include <iostream>
 #include <string>
 #include <chrono>
+#include <stdexcept>
 
 #include <libgen.h>
 
@@ -25,7 +27,7 @@ int main(int argc, char *argv[])
     g_program_name = basename(argv[0]);
 
     // Check arguments
-    if ((argc == 1)
+    if ((argc < 3)
             || (std::string(argv[1]).compare("-h") == 0)
             || (std::string(argv[1]).compare("--help") == 0))
     {
@@ -38,7 +40,7 @@ int main(int argc, char *argv[])
 
     // Create raw command string
     std::string raw_command;
-    for (int i = 1; i < argc; i++)
+    for (int i = 2; i < argc; i++)
     {
         raw_command += '\'';
         raw_command += argv[i];
@@ -46,10 +48,28 @@ int main(int argc, char *argv[])
         raw_command += " ";
     }
 
+    // Open input file
+    std::ifstream input_file(argv[1]);
+
     // Read epsilon and parameter
     std::string epsilon, parameter;
-    std::getline(std::cin, epsilon);
-    std::getline(std::cin, parameter);
+    std::getline(input_file, epsilon);
+    std::getline(input_file, parameter);
+
+    // Check and close input file
+    if (input_file.fail())
+    {
+        std::string error_msg;
+        error_msg += "Error occurred when attempting to read ";
+        error_msg += argv[1];
+        error_msg += ". ";
+        error_msg += "Make sure that ";
+        error_msg += argv[1];
+        error_msg += " is readable and contains an epsilon value on the first line";
+        error_msg += " and a parameter value on the second line.";
+        throw std::runtime_error(error_msg);
+    }
+    input_file.close();
 
     // Create input string
     std::string input_string;
@@ -86,14 +106,21 @@ void help()
     std::string help_string;
     help_string += "Usage: ";
     help_string += g_program_name;
+    help_string += " <input file>";
     help_string += " <mpi simulator>";
     help_string += " <mpi simulator args>...";
     help_string += "\n";
 
     help_string +=
-"Read epsilon and parameter from stdin, launch mpi simulator,\n"
-"and pass epsilon and parameter as input to mpi simulator.\n"
-"Once the mpi simulator has finished, print output to stdout.\n";
+R"(
+Read epsilon and parameter from input file, launch mpi simulator,
+and pass epsilon and parameter as input to mpi simulator.
+
+The epsilon value should be on the first line of the input file.
+The parameter value should be on the second line of the input file.
+
+Once the mpi simulator has finished, print output to stdout.
+)";
 
     std::cout << help_string;
 }
