### 2019-09-11 f420481
Return with zero exit status if run with help flag
