### 2019-09-11 d260dd9
Check for help flag when exactly one argument is given
