### 2019-09-11 #1716 @ThomasPak
Hi @jmlarson1, thanks for agreeing to review this project and for your comments.  Please find my responses to points 1, 2, and 3 below.  I will get back to you on the other points as soon as possible.

1. This warning shows up when [Chaste](http://www.cs.ox.ac.uk/chaste/) is not installed, or when the environment variable `CHASTE_DIR` is not set to the Chaste source directory.  It is not an error with regards to building Pakman itself, it only informs the user that the epithelial cell growth example cannot be built and/or executed.  The wiki tutorial for this example is found [here](https://github.com/ThomasPak/pakman/wiki/Example:-epithelial-cell-growth) and it contains instructions on building the example.

   The reason for this warning is that, unlike the other two examples, the epithelial cell growth example is not a self-contained example.  It requires Chaste as a dependency, which is an open-source simulation software package for problems in physiology and biology that is too big to pull into Pakman.

   Chaste users write applications by creating a project subfolder within the Chaste source tree, writing the application code, and then running CMake on Chaste and building Chaste.  More info on that [here](https://chaste.cs.ox.ac.uk/trac/wiki/ChasteGuides/UserProjects).  Therefore, Pakman needs access to a Chaste source tree so it can add the project subfolder to the Chaste source tree, run CMake on Chaste, and build the binary for the epithelial cell growth example.

   I understand that this warning may lead users to mistakenly believe that their Pakman build is somehow compromised.  Perhaps it would make more sense to let the epithelial cell growth example be controlled by a CMake flag instead?  In that way, CMake would only build and run the epithelial cell growth example when explicitly requested to do so.  Please let me know your thoughts on this.

2. Apologies for that.  I added a few tests yesterday at the end of the day, and they ran fine on my laptop but clearly they are not very portable as they are not working for you and Travis CI is also not (entirely) successful at running them.  I tried them out on a machine from my department today and it does not work there either.

   The test was intended to test the functionality of the utility executable `utils/run-mpi-simulator`.  This executable allows a user to run an MPI simulator in the terminal as if it was a standard simulator.  See [this](https://github.com/ThomasPak/pakman/wiki/Example:-epithelial-cell-growth#mpi-simulator) for an explanation what the difference is between a standard simulator and an MPI simulator.

   Briefly, whereas a standard simulator reads an `epsilon` value and a `parameter` value from the standard input, an MPI simulator receives those values from the parent MPI process that spawned it.  Normally, that parent MPI process is Pakman.  I wrote `run-mpi-simulator` as an MPI program that reads those `epsilon` and `parameter` values from the standard input and pretends to be Pakman by spawning the given MPI simulator and relaying those values to the MPI simulator in the same way Pakman would.  The purpose of `run-mpi-simulator` is for the user to be able to test out the MPI simulator in the terminal.

   However, the MPI standard does not mention standard input.  As a result, different MPI implementations will handle standard input differently.  I had forgotten about this, but I have now corrected it in commit [0d1fa18](https://github.com/ThomasPak/pakman/commit/0d1fa18ea771746ce94ecd205fafe937002266cd), and updated the [wiki tutorial](https://github.com/ThomasPak/pakman/wiki/Example:-epithelial-cell-growth#running-pakman-with-an-mpi-simulator) accordingly.  Now, `run-mpi-simulator` will read the `epsilon` and `parameter` from an input file instead of its standard input.  This should be portable across MPI implementations and it is working on Travis CI, so please pull the latest commit and give it another try.

3. You are correct that the biased coin flip is not a good example of using Pakman.  The simulator for this example is very simple and amounts to just a few simple numpy operations.  Therefore, running a simulation takes very little time and thus the parallelisation overhead will dominate the total runtime.  Pakman was indeed designed for problems where simulation times are relatively long (on the order of multiple seconds at least), so that it would benefit from parallelisation, and that is not the case here.

   However, I deliberately chose this example to be as simple as possible for two reasons.  The biased coin example is first and foremost a guide on how to use Pakman, and introduces the general workflow of using Pakman, primarily through its [wiki tutorial](https://github.com/ThomasPak/pakman/wiki/Example:-biased-coin-flip-and-ABC-rejection).

   The second reason is that in this simple case, the analytical likelihood is well characterised so we know what the result of running Pakman should converge to.  Hence, it demonstrates that the correct results are being generated.

   Because of your comment, however, I realised that not every use would realise this.  Would you suggest that I add a paragraph to the wiki tutorial to warn the user about this?
