### 2019-09-17 #1716 @gonsie
Hi,

Do I understand pakman correctly:

> Pakman has the ability to launch ABC simulators in parallel. Using the stdout from the simulators, pakman applies an ABC  sweep, rejection, or smc scheme. The simulators (and models they are running) should be supplied by users of pakman.

It would be helpful to have some more comprehensive documentation around the pakman options. That is, it would be nice for the `--help` text to appear somewhere in the web documentation so that the user could get a sense of what arguments are required for the different masters and controllers.

Do you have any examples / performance numbers of this working on an HPC cluster, possibly with a batch scheduler (e.g., slurm)? Does pakman performance scale as node-count increases?
