### 2019-09-25 #1716 @ThomasPak
Hi @gonsie,

You have summarised it very well! The only things I would add is that the stdin of simulators is also used to pass on parameters, and that the sweep functionality is not an ABC method, it simply performs a parameter sweep.

I agree that documentation on the pakman options would be helpful. I will add a wiki page listing the pakman options similarly to the `--help` text. I am also thinking of adding a "Getting Started" section in `README.md` that would include the step of getting the `--help` text to set up the right command-line options.

I have used pakman on a HPC cluster, but I have not yet made a performance benchmark for HPC clusters. I will do so and get back to you when I have the results.
