### 2019-10-08 893a57b
Improve documentation of Python scripts
