+++ examples/sis-model/plot-sis-trace.py.in
@@ -26,7 +26,7 @@ bboxtight = args.bboxtight
 outputfile = args.outputfile
 tracefile = args.tracefile
 
-# Process trace
+# Read trace from tracefile
 t_array = []
 S_array = []
 
