### 2019-10-16 c4b132d
Improve documentation of C++ code
