### 2019-10-22 #1716 @kthyng
Hi @jmlarson1, @gonsie, and @ThomasPak — how are things going? What are the next steps for this review?
