### 2019-10-23 46c6920
Improve documentation of Bash scripts
