+++ examples/sis-model/run-cpp-abc-rejection-mpi-example.sh.in
@@ -4,6 +4,7 @@ set -euo pipefail
 # Process arguments
 usage="Usage: $0 [num_accepted_parameters (default 10)]"
 
+# Check for help flag
 if [ $# -ge 1 ]
 then
     if [ $1 == "--help" ] || [ $1 == "-h" ]
@@ -13,6 +14,7 @@ then
     fi
 fi
 
+# Set number of parameters to accept
 if [ $# == 1 ]
 then
     number_accept=$1
