+++ examples/biased-coin-flip/run-cpp-abc-smc-mpi-example.sh.in
@@ -6,6 +6,7 @@ python=@PYTHON_EXECUTABLE@
 # Process arguments
 usage="Usage: $0 [population_size (default 10)] [stdev (default 0.3)]"
 
+# Check for help flag
 if [ $# -ge 1 ]
 then
     if [ $1 == "--help" ] || [ $1 == "-h" ]
@@ -15,6 +16,7 @@ then
     fi
 fi
 
+# Set population size
 if [ $# -ge 1 ]
 then
     pop_size=$1
@@ -22,6 +24,7 @@ else
     pop_size=10
 fi
 
+# Set standard deviation of perturber
 if [ $# -ge 2 ]
 then
     stdev=$2
