+++ examples/sis-model/run-python-abc-rejection-mpi-example.sh.in
@@ -1,6 +1,7 @@
 #!/bin/bash
 set -euo pipefail
 
+# Check for Python interpreter availability
 if [ "@PYTHONINTERP_FOUND@" != "TRUE" ]
 then
     echo "Cannot run Python example because Python interpreter was not found"
@@ -12,6 +13,7 @@ python=@PYTHON_EXECUTABLE@
 # Process arguments
 usage="Usage: $0 [num_accepted_parameters (default 10)]"
 
+# Check for help flag
 if [ $# -ge 1 ]
 then
     if [ $1 == "--help" ] || [ $1 == "-h" ]
@@ -21,6 +23,7 @@ then
     fi
 fi
 
+# Set number of parameters to accept
 if [ $# == 1 ]
 then
     number_accept=$1
