### 2019-10-30 4134c10
Improve paper based on comments by @jmlarson1
I have added a paragraph summarising the state of the field better.

I have added a paragraph stating clearly what problems the software is
designed to solve and who the target audience is.
