+++ paper/paper.bib
@@ -2,7 +2,6 @@
 abstract = {Many stochastic simulation approaches for generating observations from a posterior distribution depend on knowing a likelihood function. However, for many complex probability models, such likelihoods are either impossible or computationally prohibitive to obtain. Here we present a Markov chain Monte Carlo method for generating observations from a posterior distribution without the use of likelihoods. It can also be used in frequentist applications, in particular for maximum-likelihood estimation. The approach is illustrated by an example of ancestral inference in population genetics. A number of open problems are highlighted in the discussion.},
 author = {Marjoram, Paul and Molitor, John and Plagnol, Vincent and Tavare, Simon},
 doi = {10.1073/pnas.0306899100},
-file = {:home/thomas/.local/share/data/Mendeley Ltd./Mendeley Desktop/Downloaded/Marjoram et al. - 2003 - Markov chain Monte Carlo without likelihoods.pdf:pdf},
 isbn = {0027-8424},
 issn = {0027-8424},
 journal = {Proc. Natl. Acad. Sci. U. S. A.},
@@ -22,7 +21,6 @@ arxivId = {0901.1925},
 author = {Toni, Tina and Welch, David and Strelkowa, Natalja and Ipsen, Andreas and Stumpf, Michael P.H.},
 doi = {10.1098/rsif.2008.0172},
 eprint = {0901.1925},
-file = {:home/thomas/.local/share/data/Mendeley Ltd./Mendeley Desktop/Downloaded/Toni et al. - 2009 - Approximate Bayesian computation scheme for parameter inference and model selection in dynamical systems.pdf:pdf},
 isbn = {1742-5689},
 issn = {17425662},
 journal = {J. R. Soc. Interface},
@@ -38,7 +36,6 @@ year = {2009}
 abstract = {Mechanistic understanding of multi-scale biological processes, such as cell proliferation in a changing biological tissue, is readily facilitated by computational models. While tools exist to construct and simulate multi-scale models, the statistical inference of the unknown model parameters remains an open problem. Here, we present and benchmark a parallel approximate Bayesian computation sequential Monte Carlo (pABC SMC) algorithm, tailored for high-performance computing clusters. pABC SMC is fully automated and returns reliable parameter estimates and confidence intervals. By running the pABC SMC algorithm for ∼106 hr, we parameterize multi-scale models that accurately describe quantitative growth curves and histological data obtained in vivo from individual tumor spheroid growth in media droplets. The models capture the hybrid deterministic-stochastic behaviors of 105–106 of cells growing in a 3D dynamically changing nutrient environment. The pABC SMC algorithm reliably converges to a consistent set of parameters. Our study demonstrates a proof of principle for robust, data-driven modeling of multi-scale biological systems and the feasibility of multi-scale model parameterization through statistical inference.},
 author = {Jagiella, Nick and Rickert, Dennis and Theis, Fabian J. and Hasenauer, Jan},
 doi = {10.1016/j.cels.2016.12.002},
-file = {:home/thomas/.local/share/data/Mendeley Ltd./Mendeley Desktop/Downloaded/Jagiella et al. - 2017 - Parallelization and High-Performance Computing Enables Automated Statistical Inference of Multi-scale Models.pdf:pdf},
 issn = {24054720},
 journal = {Cell Syst.},
 keywords = {Bayesian parameter estimation,approximate Bayesian computation,high-performance computing,model-based data integration,multi-scale modeling,statistical inference,tumor spheroids},
@@ -75,7 +72,6 @@ year = {2011}
 @book{MPI2015,
 address = {Knoxville},
 author = {{Message Passing Interface Forum}},
[…]
