### 2019-11-04 #1716 @ThomasPak
Hi @gonsie

As per your suggestion, I have added a Wiki page documenting the command-line interface of Pakman here: https://github.com/ThomasPak/pakman/wiki/Command-line-interface

I have performed a benchmark on my University's HPC cluster. The cluster uses Slurm for job scheduling and I performed the test on 16 nodes with 16 CPU cores each. The CPUs were all Intel Xeon E5-2650 processors. Briefly, I sent a job to the cluster with Pakman to run 4096 simulations of `scaling/scaling-simulator.sh` with varying numbers of CPU cores used. The numbers of CPU cores range from 1 to 256 in powers of 2. The resulting computation times are given below.

Number of CPU cores | Computation time
-|-
1|3h 35m 50.286s
2|1h 49m 06.142s
4|57m 08.339s
8|30m 02.199s
16|15m 43.652s
32|10m 28.441s
64|3m 53.066s
128|2m 26.269s
256|1m 19.147s

Moreover, I have plotted the speedup and efficiency (defined with respect to computation times under ideal linear scaling) using `scaling/plot-scaling.py` below.

![speedup](https://user-images.githubusercontent.com/12231744/68140777-4ac76d80-ff24-11e9-8386-adede555fe8a.png)

![efficiency](https://user-images.githubusercontent.com/12231744/68140765-44d18c80-ff24-11e9-98d3-972194461e95.png)

The computation times scale roughly linear, although there is a drop in efficiency at higher core counts. Part of this can be attributed to overhead involved with to message passing, while part of it may be because of inefficiencies in the code. Since the average duration of a single simulation in this example is roughly three seconds, the parallel overhead would have a smaller impact for longer-running simulations.

Overall, I believe that these results demonstrate that Pakman scales fairly well with increased computational resources.
