### 2019-11-06 #1716 @gonsie
Thanks @ThomasPak. Maybe you want to document the performance graphs/details to the article? Or maybe to a wiki page? Otherwise, I think it looks good.
