### 2019-11-06 #1716 @jmlarson1
I believe the checklist requirements have been satisfied.
