### 2019-11-08 #1716 @ThomasPak
@gonsie I have put the results of this performance test on the Pakman wiki because I may add more tests in the future. Please have a look at https://github.com/ThomasPak/pakman/wiki/Benchmark%3A-HPC-cluster
