### 2019-11-22 #1716 @jedbrown
I see the checklists are complete; thanks. @gonsie @jmlarson1 Please let me know if you have any further concerns.
