### 2019-12-01 #1716 @ThomasPak
> I see the checklists are complete; thanks. @gonsie @jmlarson1 Please let me know if you have any
further concerns.

Hi @gonsie @jmlarson1, have you had time to look into this?
