### 2019-12-25 #1716 @jedbrown
Thanks for your patience.

> Pakman (**P**arallel **A**B**C** **Man**ager)

Although I know what you mean, this just reads like a typo. Perhaps drop the bold and just write something like the following?

> Pakman is a parallel ABC manager that is designed to be ...

Is MPI-3.1 really needed, or would an earlier standard suffice?

I'm curious about failure. Is a job allowed to fail? What if it exceeds memory capacity (e.g., due to adaptive methods or large Krylov space)? The interaction between MPI and `fork`/`exec` is delicate, and I don't know if the present usage is allowed by the standard or just happens to work with the implementations you have tried. Could you clarify how `MPI_Comm_spawn` versus `fork`/`exec` are used? If I'm following context correctly, it would be useful for the `MPI_Info` argument to be user-tunable, because it may be needed to interact properly with a resource manager.
