### 2020-01-13 f749ffe
Remove caption link breaking pdf compilation
