### 2020-01-13 33df131
Remove bold in definition of Pakman
