### 2020-01-13 #1716 @ThomasPak
Hi @jedbrown,

I have changed the text to "Pakman is a parallel ABC manager that is designed to be ..." as you suggested.

I have not tested Pakman with older MPI versions because I have only used the most recent MPI implementations. Would you recommend doing so?

### Failure

Simulation jobs are allowed to exit with a nonzero return code. The user can decide whether to ignore these errors with the command-line option `--ignore-errors`. However, if the simulator exits abnormally, for instance because of a memory error, an error will be thrown and Pakman will quit.

### `fork`/`exec`

I looked through the MPI standard and there is no mention of `fork`/`exec`. Thus, there are no guarantees on whether `fork`/`exec` will work as intended. However, it does work with the OpenMPI and MPICH implementations that I have tried so far. In the case of OpenMPI, forking is supported explicitly, and a warning is generated when it is unsafe to fork (see https://www.open-mpi.org/faq/?category=tuning#fork-warning). On the other hand, MPICH does not mention forking anywhere in its documentation, but in practice I have had no issues forking with an MPICH implementation. I realise that this makes Pakman less portable, so would you suggest adding a portability note about this in README.md?

### `MPI_Comm_spawn` versus `fork`/`exec`

These two spawning methods are mutually exclusive. I initially only implemented job spawning using `fork`/`exec`, but then I discovered that this is not compatible with simulators that use MPI themselves. Hence, I implemented an alternative method of job spawning using `MPI_Comm_spawn`. More information on this can be found on https://codedocs.xyz/ThomasPak/pakman/mpi-simulator.html.

### `MPI_Info`

Thanks for pointing this out, you are absolutely right. In fact, I have already come across a situation where it was necessary to write to an `MPI_Info` object. In particular, MPICH by default does not spawn new MPI processes (when using `MPI_Comm_spawn`) on the same host as the parent MPI process. Instead, it spawns all new MPI processes on the host of the root MPI process. Therefore, the command-line option `--force-host-spawn` ensures that the `host` key in `MPI_Info` is set to the hostname of the MPI process that is calling `MPI_Comm_spawn` to ensure that the child MPI process is spawned on the same host as the parent MPI process.

As you said however, it makes sense to give the user the flexibility to specify arbitrary key-value pairs in `MPI_Info`. I will implement this functionality and get back to you so you can review it.
