### 2020-01-13 #1716 @ThomasPak
@jedbrown

In addition, with regards to the pull request you created, I noticed that the links to the GitHub wiki inside the figure captions were breaking the pdf compilation. I removed them for now so that it compiles correctly, but if there is an alternative way to do it that does not break the pdf compilation, I am happy to put the links back in.
