### 2020-01-15 0510e25
Add function to parse key-value pairs from string
