### 2020-01-15 9cc1a98
Add option to tune MPI_Info to MPI_Comm_spawn
