### 2020-01-15 02ce1f2
Free MPI_Info object
