### 2020-01-15 #1716 @ThomasPak
@jedbrown

The MPI_Info object passed to MPI_Comm_spawn can now be tuned by a command-line option to Pakman. Please see https://github.com/ThomasPak/pakman/commit/9cc1a98c22924bbb160ff74e3c2a883e0cdf43bb
