### 2020-02-27 #1716 @jedbrown
@ThomasPak I apologize for my greatly delayed response here. With regard to the experiment in Figure 2, is there a set of parameters that would acceptably reproduce the analytic result? Does Pakman have any functionality to inform the user when their choices are insufficient and thus can't be trusted? Also, is there a reason you to use `\langle 75, 70, ..., 35 \rangle` as notation for what I understand to just be a list (in which case braces `\{75, ..., 35\}` are more common)?

Finally, please update the bibliography to use full journal names instead of the (PubMed?) abbreviations like `Stat. Appl. Genet. Mol. Biol.`. If that renders alright, I think we'll be ready to archive.
