### 2020-02-27 b13b7c4
Use full journal names
