+++ paper/paper.bib
@@ -4,7 +4,7 @@ author = {Marjoram, Paul and Molitor, John and Plagnol, Vincent and Tavare, Simo
 doi = {10.1073/pnas.0306899100},
 isbn = {0027-8424},
 issn = {0027-8424},
-journal = {Proc. Natl. Acad. Sci. U. S. A.},
+journal = {Proceedings of the National Academy of Sciences of the United States of America},
 keywords = {Algorithms,Biological,Biological Evolution,Computer Simulation,DNA,DNA: genetics,Genetics,Humans,Likelihood Functions,Markov Chains,Mitochondrial,Mitochondrial: genetics,Models,Monte Carlo Method,Population,Stochastic Processes},
 number = {26},
 pages = {15324--15328},
@@ -23,7 +23,7 @@ doi = {10.1098/rsif.2008.0172},
 eprint = {0901.1925},
 isbn = {1742-5689},
 issn = {17425662},
-journal = {J. R. Soc. Interface},
+journal = {Journal of the Royal Society Interface},
 keywords = {Bayesian model selection,Dynamical systems,Parameter estimation,Sequential Monte Carlo,Sequential importance sampling,Sloppy parameters},
 number = {31},
 pages = {187--202},
@@ -37,7 +37,7 @@ abstract = {Mechanistic understanding of multi-scale biological processes, such 
 author = {Jagiella, Nick and Rickert, Dennis and Theis, Fabian J. and Hasenauer, Jan},
 doi = {10.1016/j.cels.2016.12.002},
 issn = {24054720},
-journal = {Cell Syst.},
+journal = {Cell Systems},
 keywords = {Bayesian parameter estimation,approximate Bayesian computation,high-performance computing,model-based data integration,multi-scale modeling,statistical inference,tumor spheroids},
 number = {2},
 pages = {194--206},
@@ -83,7 +83,7 @@ author = {Mirams, Gary R. and Arthurs, Christopher J. and Bernabeu, Miguel O. an
 doi = {10.1371/journal.pcbi.1002970},
 editor = {Prlic, Andreas},
 issn = {1553-7358},
-journal = {PLoS Comput. Biol.},
+journal = {PLoS Computational Biology},
 number = {3},
 pages = {e1002970},
 publisher = {Public Library of Science},
@@ -98,7 +98,7 @@ author = {Osborne, James M. and Fletcher, Alexander G. and Pitt-Francis, Joe M. 
 doi = {10.1371/journal.pcbi.1005387},
 isbn = {1111111111},
 issn = {15537358},
-journal = {PLoS Comput. Biol.},
+journal = {PLoS Computational Biology},
 number = {2},
 pages = {e1005387},
 pmid = {28192427},
@@ -124,7 +124,7 @@ abstract = {Approximate Bayesian computation (ABC) has gained popularity over th
 author = {Filippi, Sarah and Barnes, Chris P. and Cornebise, Julien and Stumpf, Michael P.H.},
 doi = {10.1515/sagmb-2012-0069},
 issn = {15446115},
-journal = {Stat. Appl. Genet. Mol. Biol.},
+journal = {Statistical Applications in Genetics and Molecular Biology},
 keywords = {Adaptive kernels,Bayesian parameter inference,Dynamical systems,Kullback-Leibler,Likelihood-free,Sequential Monte Carlo},
 number = {1},
 pages = {87--107},
@@ -137,7 +137,7 @@ abstract = {Approximate Bayesian Computational (ABC) methods (or likelihood-free
 author = {Baragatti, Me{\"{i}}li and Grimaud, Agn{\`{e}}s and Pommeret, Denys},
 doi = {10.1007/s11222-012-9328-6},
 issn = {09603174},
-journal = {Stat. Comput.},
[…]
