### 2020-03-04 a73fdca
Remove editor entries
