+++ paper/paper.bib
@@ -50,7 +50,6 @@ year = {2017}
 @article{Klinger2018,
 author = {Klinger, Emmanuel and Rickert, Dennis and Hasenauer, Jan},
 doi = {10.1093/bioinformatics/bty361},
-editor = {Stegle, Oliver},
 issn = {1367-4803},
 journal = {Bioinformatics},
 month = {oct},
@@ -81,7 +80,6 @@ year = {2015}
 abstract = {Chaste — Cancer, Heart And Soft Tissue Environment — is an open source C++ library for the computational simulation of mathematical models developed for physiology and biology. Code development has been driven by two initial applications: cardiac electrophysiology and cancer development. A large number of cardiac electrophysiology studies have been enabled and performed, including high-performance computational investigations of defibrillation on realistic human cardiac geometries. New models for the initiation and growth of tumours have been developed. In particular, cell-based simulations have provided novel insight into the role of stem cells in the colorectal crypt. Chaste is constantly evolving and is now being applied to a far wider range of problems. The code provides modules for handling common scientific computing components, such as meshes and solvers for ordinary and partial differential equations (ODEs/PDEs). Re-use of these components avoids the need for researchers to ‘re-invent the wheel' with each new project, accelerating the rate of progress in new applications. Chaste is developed using industrially-derived techniques, in particular test-driven development, to ensure code quality, re-use and reliability. In this article we provide examples that illustrate the types of problems Chaste can be used to solve, which can be run on a desktop computer. We highlight some scientific studies that have used or are using Chaste, and the insights they have provided. The source code, both for specific releases and the development version, is available to download under an open source Berkeley Software Distribution (BSD) licence at http://www.cs.ox.ac.uk/chaste, together with details of a mailing list and links to documentation and tutorials.},
 author = {Mirams, Gary R. and Arthurs, Christopher J. and Bernabeu, Miguel O. and Bordas, Rafel and Cooper, Jonathan and Corrias, Alberto and Davit, Yohan and Dunn, Sara-Jane and Fletcher, Alexander G. and Harvey, Daniel G. and Marsh, Megan E. and Osborne, James M. and Pathmanathan, Pras and Pitt-Francis, Joe and Southern, James and Zemzemi, Nejib and Gavaghan, David J.},
 doi = {10.1371/journal.pcbi.1002970},
-editor = {Prlic, Andreas},
 issn = {1553-7358},
 journal = {PLoS Computational Biology},
 number = {3},
@@ -215,7 +213,6 @@ arxivId = {1806.08320},
 author = {Kousathanas, Athanasios and Duchen, Pablo and Wegmann, Daniel},
 booktitle = {Handbook of Approximate {Bayesian} Computation},
 chapter = {13},
-editor = {Sisson, Scott A. and Fan, Yanan and Beaumont, Mark},
 eprint = {1806.08320},
 pages = {369--414},
 publisher = {Chapman {and} Hall/CRC},
