### 2020-03-04 #1716 @ThomasPak
@jedbrown No worries!

> With regard to the experiment in Figure 2, is there a set of parameters that would acceptably reproduce the analytic result?

Yes. The reason why the computational results are different from the analytical results is that the tolerance series used ends with a tolerance of 35. Whenever the target tolerance in the ABC SMC is nonzero, the computed posterior is only an approximation of the posterior. If the target tolerance was zero, however, the computed posterior would reproduce the analytical result.

Thanks for asking this question; because of your question I discovered an error in the caption of Figure 2. In particular, I wrote that the reasons for the discrepancy were insufficient summary statistics and a nonzero tolerance. However, no summary statistics were used, so the only reason for the discrepancy is a nonzero tolerance.

>  Does Pakman have any functionality to inform the user when their choices are insufficient and thus can't be trusted?

Pakman currently does not have this functionality because there is no easy and straightforward way to check the quality of a computed posterior. The most general method is to use posterior predictive checks, which involves generating 'fake' data from the fitted model and comparing it to the real data. We did not include this in the scope of Pakman, although we do not rule out incorporating this functionality in the future.

> Also, is there a reason you to use `\langle 75, 70, ..., 35 \rangle` as notation for what I understand to just be a list (in which case braces `\{75, ..., 35\}` are more common)?

We used angle brackets to stress that the order of elements is important. If we used braces, it could be mistaken for a set instead of an ordered list.

> Finally, please update the bibliography to use full journal names instead of the (PubMed?) abbreviations like `Stat. Appl. Genet. Mol. Biol.`. If that renders alright, I think we'll be ready to archive.

I have made these changes.

Thanks for your comments and please let me know if you are happy to proceed with archiving.
