### 2020-03-04 #1716 @jedbrown
Thanks for your explanation and updates. At this point, here are the next steps:
- [ ] Make a tagged release of your software (`v1.0.1` or whatever is appropriate), and list the version tag of the archived version here. [Annotated tags](https://git-scm.com/book/en/v2/Git-Basics-Tagging#_annotated_tags) are preferred.
- [ ] Archive the reviewed software in Zenodo or a similar service (e.g. figshare, an institutional repository)
- [ ] Check the archival deposit (e.g., in Zenodo) has the correct metadata, this includes the title (should match the paper title) and author list (make sure the list is correct and people who only made a small fix are not on it); you may also add the authors' ORCID.
