### 2020-03-06 #1716 @ThomasPak
The paper PDF and Crossref deposit XML look good to me :thumbsup:.
