### 2020-03-06 #1716 @Kevin-Mattheus-Moerman
@ThomasPak I'm helping with the final processing for your paper. Can you check the point below.
- [ ] The following sentence needs rephrasing, please remove the words "embarrassingly": `Fortunately, some ABC algorithms have a simulation workload that is embarrassingly parallel, or proceed through a sequence of embarrassingly parallel simulation workloads.`,
