### 2020-03-07 #1716 @jedbrown
@Kevin-Mattheus-Moerman Other wordings are possible, but it as a [widely used](https://scholar.google.com/scholar?hl=en&q=embarrassingly%20parallel&as_sdt=1%2C6&as_sdtp=) [technical term](https://en.wikipedia.org/wiki/Embarrassingly_parallel), and wouldn't have the same meaning if you simply cut "embarrassingly".
